Moved POINTS_DATABASE into a constants file
