+++ src/ui/setup/dynamic/devtools.ts
@@ -32,7 +32,7 @@ import { addEventListener, initSocket, client as protocolClient } from "protocol
 import { ThreadFront } from "protocol/thread";
 import { assert } from "protocol/utils";
 import { CONSOLE_SETTINGS_DATABASE } from "replay-next/src/contexts/ConsoleFiltersContext";
-import { POINTS_DATABASE } from "replay-next/src/contexts/points/PointsContext";
+import { POINTS_DATABASE } from "replay-next/src/contexts/points/constants";
 import { IDBOptions } from "replay-next/src/hooks/useIndexedDB";
 import {
   getCachedObject,
